// src/shared/dataprovider/api/model/clientCOFundsBalanceDTO.ts
/**
 * DynastyGateway API
 * API Connections to DPay
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ClientCOFundsBalanceDTO { 
    ClientId?: number;
    Name?: string;
    AvailableBalance?: number;
}